melhorias de do codigo

+++ horas_estagio/src/main/java/com/registo/horas_estagio/dto/response/RegisterResponse.java
@@ -10,6 +10,6 @@ public record RegisterResponse(
         String descricao,
         LocalDateTime dataInicio,
         LocalDateTime dataFim,
-     int horasTrabalhadas
+        double horasTrabalhadas
 ) {
 }
